classes reorganization
