+++ src/main/java/tkcy/simpleaddon/api/recipes/helpers/RecipeRemovalHelper.java
@@ -1,4 +1,4 @@
-package tkcy.simpleaddon.api.recipes;
+package tkcy.simpleaddon.api.recipes.helpers;
 
 import javax.annotation.Nonnull;
