+++ src/main/java/tkcy/simpleaddon/api/recipes/logic/ToolRecipeLogic.java
@@ -21,8 +21,8 @@ import gregtech.api.util.GTTransferUtils;
 import gregtech.api.util.GTUtility;
 
 import tkcy.simpleaddon.api.metatileentities.ToolLogicMetaTileEntity;
-import tkcy.simpleaddon.api.recipes.RecipeSearchHelpers;
 import tkcy.simpleaddon.api.recipes.builders.ToolRecipeBuilder;
+import tkcy.simpleaddon.api.recipes.helpers.RecipeSearchHelpers;
 import tkcy.simpleaddon.api.recipes.properties.ToolProperty;
 import tkcy.simpleaddon.api.recipes.properties.ToolUsesProperty;
 import tkcy.simpleaddon.modules.NBTHelpers;
