+++ src/main/java/tkcy/simpleaddon/api/recipes/recipemaps/ToolRecipeMap.java
@@ -1,4 +1,4 @@
-package tkcy.simpleaddon.api.recipes;
+package tkcy.simpleaddon.api.recipes.recipemaps;
 
 import net.minecraftforge.items.IItemHandlerModifiable;
