+++ src/main/java/tkcy/simpleaddon/common/metatileentities/TKCYSAMetaTileEntities.java
@@ -9,7 +9,7 @@ import gregtech.api.util.GTUtility;
 import gregtech.client.renderer.texture.Textures;
 
 import tkcy.simpleaddon.api.metatileentities.ToolLogicMetaTileEntity;
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 import tkcy.simpleaddon.api.render.TKCYSATextures;
 import tkcy.simpleaddon.api.utils.TKCYSAUtil;
 import tkcy.simpleaddon.common.metatileentities.electric.*;
