+++ src/main/java/tkcy/simpleaddon/common/metatileentities/electric/AssemblingMachine.java
@@ -20,7 +20,7 @@ import gregtech.api.pattern.FactoryBlockPattern;
 import gregtech.client.renderer.ICubeRenderer;
 import gregtech.client.renderer.texture.Textures;
 
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 
 public class AssemblingMachine extends RecipeMapMultiblockController {
