+++ src/main/java/tkcy/simpleaddon/common/metatileentities/electric/Dryer.java
@@ -19,7 +19,7 @@ import gregtech.client.renderer.texture.Textures;
 import gregtech.common.blocks.BlockMachineCasing;
 import gregtech.common.blocks.MetaBlocks;
 
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 
 public class Dryer extends RecipeMapMultiblockController {
