+++ src/main/java/tkcy/simpleaddon/common/metatileentities/electric/Electrolyzer.java
@@ -19,7 +19,7 @@ import gregtech.api.pattern.FactoryBlockPattern;
 import gregtech.client.renderer.ICubeRenderer;
 import gregtech.client.renderer.texture.Textures;
 
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 
 public class Electrolyzer extends RecipeMapMultiblockController {
