+++ src/main/java/tkcy/simpleaddon/common/metatileentities/multiprimitive/FluidPrimitiveBlastFurnace.java
@@ -17,7 +17,7 @@ import gregtech.client.renderer.ICubeRenderer;
 import gregtech.client.renderer.texture.Textures;
 
 import tkcy.simpleaddon.api.machines.NoEnergyMultiController;
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 
 public class FluidPrimitiveBlastFurnace extends NoEnergyMultiController {
