+++ src/main/java/tkcy/simpleaddon/common/metatileentities/multiprimitive/PrimitiveRoastingOven.java
@@ -18,7 +18,7 @@ import gregtech.client.renderer.ICubeRenderer;
 import gregtech.client.renderer.texture.Textures;
 
 import tkcy.simpleaddon.api.machines.NoEnergyMultiController;
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 
 public class PrimitiveRoastingOven extends NoEnergyMultiController {
