+++ src/main/java/tkcy/simpleaddon/common/metatileentities/steam/SteamMelter.java
@@ -25,7 +25,7 @@ import gregtech.client.particle.VanillaParticleEffects;
 import gregtech.client.renderer.texture.Textures;
 import gregtech.client.renderer.texture.cube.SimpleSidedCubeRenderer;
 
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 
 public class SteamMelter extends SteamMetaTileEntity {
