+++ src/main/java/tkcy/simpleaddon/loaders/recipe/alloys/AlloyingRecipes.java
@@ -16,7 +16,7 @@ import gregtech.api.unification.material.Material;
 import gregtech.api.unification.ore.OrePrefix;
 import gregtech.api.unification.stack.MaterialStack;
 
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 import tkcy.simpleaddon.api.utils.MaterialHelper;
 
 public class AlloyingRecipes {
