+++ src/main/java/tkcy/simpleaddon/loaders/recipe/chains/metals/AluminiumChain.java
@@ -11,7 +11,7 @@ import static tkcy.simpleaddon.api.unification.ore.TKCYSAOrePrefix.cathode;
 import gregtech.api.GTValues;
 import gregtech.api.recipes.RecipeMaps;
 
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 
 public class AluminiumChain {
