+++ src/main/java/tkcy/simpleaddon/loaders/recipe/chains/metals/FluorineChain.java
@@ -8,7 +8,7 @@ import static gregtech.api.unification.ore.OrePrefix.dust;
 import static gregtech.api.unification.ore.OrePrefix.dustTiny;
 import static tkcy.simpleaddon.api.unification.materials.TKCYSAMaterials.*;
 
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 import tkcy.simpleaddon.api.unification.ore.TKCYSAOrePrefix;
 
 public class FluorineChain {
