+++ src/main/java/tkcy/simpleaddon/loaders/recipe/handlers/PrimitiveCastingHandler.java
@@ -7,7 +7,7 @@ import gregtech.api.unification.material.Material;
 import gregtech.api.unification.ore.OrePrefix;
 
 import tkcy.simpleaddon.api.recipes.CastingInfo;
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 import tkcy.simpleaddon.api.utils.MaterialHelper;
 import tkcy.simpleaddon.api.utils.TKCYSALog;
