+++ src/main/java/tkcy/simpleaddon/loaders/recipe/handlers/Roasting.java
@@ -9,8 +9,8 @@ import static tkcy.simpleaddon.api.unification.materials.TKCYSAMaterials.*;
 import gregtech.api.recipes.RecipeMaps;
 import gregtech.api.unification.material.Material;
 
-import tkcy.simpleaddon.api.recipes.RecipeRemovalHelper;
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.helpers.RecipeRemovalHelper;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 
 public class Roasting {
