+++ src/main/java/tkcy/simpleaddon/loaders/recipe/handlers/harderstuff/HarderCoilsRecipes.java
@@ -13,7 +13,7 @@ import gregtech.common.blocks.BlockWireCoil;
 import gregtech.common.blocks.MetaBlocks;
 import gregtech.common.items.MetaItems;
 
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 
 public class HarderCoilsRecipes {
