+++ src/main/java/tkcy/simpleaddon/loaders/recipe/parts/PrimitivePartsHandler.java
@@ -3,7 +3,7 @@ package tkcy.simpleaddon.loaders.recipe.parts;
 import gregtech.api.unification.material.Materials;
 import gregtech.api.unification.ore.OrePrefix;
 
-import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 import tkcy.simpleaddon.modules.ToolsModule;
 
 public class PrimitivePartsHandler {
